+++ demo/cypress/index.ts
@@ -9,4 +9,4 @@ export { VStepper } from './VStepper';
 export { DatePicker } from './DatePicker';
 export { DateRange } from './DateRange';
 export { DropDownSelect } from './DropDownSelect';
-
+export { Wizard } from './Wizard';