+++ leda/components/Wizard/theme.ts
@@ -1,0 +1,13 @@
+/* eslint-disable key-spacing */
+export const defaultWizardTheme = {
+  label:                    'bottom',
+  wizardStep:               'step',
+  wizardItemActive:         'active',
+  wizardItemFirst:          'first',
+  wizardItemLast:           'last',
+  wizardItemProgress:       'progress',
+  wizardItemSuccess:        'success',
+  wizardLine:               'line',
+  wrapper:                  'wizard',
+};
+  /* eslint-enable key-spacing */