+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [0.31.0] - 2020-09-28
+- Wizard: Added new Wizard component
 
 ## [0.30.0] - 2020-09-14